Refactor `efibootmgr` examples with tabs
Updates the `systemd-boot` section to wrap `efibootmgr` examples in tab blocks, allowing users to toggle between using the recommended `--unicode` switch and running the command without it.

This change is introduced because firmware compliance with the UEFI Unicode specification can vary, leading to issues if the wrong setting is used.

The affected commands are:

- `efibootmgr` (viewing entries)
- `efibootmgr | grep...` (identifying new entry)
- `efibootmgr --bootorder` (modifying order)
- `efibootmgr --delete-bootnum` (removing stale entries)

A new `!!! info inline end` block is added to explain the role of the `--unicode` switch.